Add a workflow to initialize the branches
The idea is that branches (whose name reflect the targeted CPU
architecture) hold the respective state of the Pacman repository.

Currently, Git for Windows' Pacman repository is hosted on Azure Blobs,
in my personal Azure account. This is not ideal, an open source
maintainer should always be ready to step away.

So let's download the files that are currently referenced by the
database at Git for Windows' repository, initialize these branches,
to test whether we could easily switch to using

  [git-for-windows]
  Server = https://github.com/git-for-windows/pacman-repo/raw/refs/heads/x86_64

instead of

  [git-for-windows]
  Server = https://wingit.blob.core.windows.net/x86-64

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>