Add the list of Azure resources
Git for Windows' Pacman repository is currently hosted on Azure Blobs,
in my personal account. To disentangle Git for Windows from my person,
I would like to move the contents to a new GitHub repository,
https://github.com/git-for-windows/pacman-repo.

As a first step, I created a workflow at
https://github.com/git-for-windows/pacman-repo/commits/initialize-branches
that downloaded the current tip packages and databases from the Pacman
repository and initialized the `x86_64`, `i686`, and `aarch64` branches
to reflect that state. The idea being that editing `/etc/pacman.conf` as
following would do the same thing as Git for Windows does currently,
except that the packages would be served from GitHub instead of my
personal Azure Blob Storage account:

  [git-for-windows]
  Server = https://github.com/git-for-windows/pacman-repo/raw/refs/heads/x86_64

However, this just reflects the _current_ state as of time of writing,
which will soon be outdated.

In particular, it does _not_ reflect the historical state. There have
been tons of updates to the Pacman repository of Git for Windows, and
these updates have been everything but transparent, decipherable nor
easily reasoned about.

Let's start by adding the list of files that had been uploaded to said
Azure Blobs Storage account. These were obtained using command-lines
like this one:

    $ az storage blob list \
        -c x86-64 --account-name wingit >x86-64-list.txt

This was done for the x86-64, i686, aarch64, and sources containers.

The next step will be to coalesce those into a list of files aggregated
by (inexact) upload time. This will serve as a starting point to define
which files need to be uploaded to GitHub releases in the `pacman-repo`
repository to address above-mentioned transparency challenges.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>