Add a script to generate the list of releases
The idea is to map the current Git for Windows' Pacman repository to a
set of releases in https://github.com/git-for-windows/pacman-repo.

Sadly, this set of releases can _not_ serve as the new Pacman repository
for Git for Windows, because Pacman repositories have a flat directory
structure: the database file needs to live in the same "directory" as
the packages. In other words, if the database lives in

  https://github.com/git-for-windows/pacman-repo/releases/download/latest-git-for-windows.db

then the packages need to live in the same GitHub release. I do not want
that, though, as it would be wasteful to replicate the package files
every time the Pacman repository is updated, as opposed to uploading
only the files for the updated package to the releases.

However, for transparency reasons, it is a good idea to upload the
package files (even historical versions) to GitHub releases. That allows
investigators (or: interested Git for Windows users) to inspect what
package versions had been used in Git for Windows at various stages.

To prepare for that, here is a node.js script that parses the
just-downloaded JSON reflecting what has been uploaded at various times
to Git for Windows' Pacman repository. It then outputs a list of
timestamps and sets of files that should be turned into GitHub releases
and assets.

This will finally disentangle Git for Windows' Pacman repository from my
personal Azure account, as it should have been in the first place.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>