Add a workflow to initialize the releases
Git for Windows' Pacman repository is too opaque. Let's first mirror it
to GitHub releases at https://github.com/git-for-windows/pacman-repo.

The idea is to eventually switch completely to that transparent
process of uploading new packages to GitHub releases in that repository.
And after that, updating its x86_64, aarch64 and i686 branches
(representing Git for Windows' new Pacman repository) to reflect the
updated packages.

Since GitHub Actions' debugging facilities are totally insufficient,
let's first try this with the initial release and then stop. In the
unlikely event that it works, I will then proceed to handle the
remaining releases.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>