Use a more natural name for the non-MSYS database
The package databases of Git for Windows come in two flavors: one that
contains all the packages for a specific CPU architecture, and another
one that contains the subset of packages that excludes the MSYS ones.

The reason for this split is: any given MSYS2 installation can install
MSYS packages only for the CPU architecture whose MSYS2 runtime is
installed, while the same is not true for MINGW packages.

For example, it is quite possible for an MSYS2 setup on Windows/ARM64
to contain x86_64, aarch64 _and_ i686 flavors of `mingw-w64-curl`, but
the `curl` package can be installed only once, matching the CPU
architecture of `/usr/bin/msys-2.0.dll`.

Therefore, historically Git for Windows provided four package databases
in its Pacman repositories:

- `x86-64/git-for-windows.db` (MSYS + MINGW),
- `x86-64/git-for-windows-mingw64.db` (MINGW only),
- `i686/git-for-windows.db` (MSYS + MINGW), and
- `i686/git-for-windows-mingw32.db` (MINGW only).

Then, `git-sdk-64` would use `x86-64/git-for-windows.db` and
`i686/git-for-windows-mingw32.db`, while `git-sdk-32` would use
`i686/git-for-windows.db` and `x86-64/git-for-windows-mingw64.db`.

Later on, when Git for Windows started the first timid steps to support
Windows/ARM64, the same pair of databases was created for the `aarch64`
packages, with two twists:

- The `aarch64/git-for-windows.db` database does not even contain any
  MSYS packages because we do not have any MSYS2 runtime for
  Windows/ARM64 yet! Therefore, it is identical to the MINGW-only
  database.
- The MINGW-only database did not follow the previous naming scheme
  where the down-cased `MSYSTEM` value was used as a suffix to discern
  the MINGW-only from the MSYS+MINGW database.

That latter point was a mistake, and it matters now that I want to add
all files associated with a Pacman repository update to a GitHub
Release in `git-for-windows/pacman-repo`, as the MSYS+MINGW database
names clash. The most natural way would be to add the CPU architecture
as suffix, but that is now unfortunately taken by the MINGW-only
database.

However, since I am shaking up things anyway (requiring an update to
`/etc/pacman.conf` to move to the new Pacman repository location), I
might as well rename the database to something more natural _now_.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>