database: add the CPU architecture as a suffix
To allow for uploading all the files associated with a Pacman deployment
to a single GitHub Release in `git-for-windows/pacman-repo`, including
the package database files for all CPU architectures, the latters' names
must be unique.

This is currently not the case, they MSYS+MINGW variants of the package
database is named `git-for-windows.db` for all of them.

Let's fix that by adding the CPU architecture as a suffix to the
database's filename.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>