Replace the README contents with some actually useful information
One week ago, I started the experiment whether it would be possible to
migrate Git for Windows' Pacman repository from Azure Blobs to a GitHub
repository.

After a couple of discussions and a few pivots of the design, I opened
and merged a couple of PRs with the effect to dual-deploy Git for
Windows' packages both to Azure Blobs as well as to `pacman-repo`.

Now that things have shaken out, I am confident that this is a good
approach and am ready to continue with the migration, one important part
of which is to describe the architecture. This repository's README is
as fine a place as any to do just that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>